Add files via upload
This code reports the intermediate iterates of the Newton-Raphson inverse kinematics.  A created function solves the inverse kinematics of the UR5 industrial robot. The convergence of the iterations was visualized using the CoppeliaSim robot simulator.
This project is offered by the course- Modern Robotics: Mechanics, Planning, and Control offered by Northwestern University on Coursera.